#118: moving database -> demo_database folder, and moving examples
