+++ demo_database/src/main/postgres/integration/V1.2.12__insert_other_types.sql
