+++ demo_database/src/main/postgres/integration/V1.2.4__other_types.ddl
