--- examples/src/main/resources/application.conf
@@ -1,12 +1,0 @@
-menasdb = {
-  connectionPool = "HikariCP" //use HikariCP for our connection pool
-  dataSourceClass = "org.postgresql.ds.PGSimpleDataSource" //Simple datasource with no connection pooling. The connection pool has already been specified with HikariCP.
-  properties = {
-    serverName = "localhost"
-    portNumber = "5432"
-    databaseName = "menas_db"
-    user = "menas"
-    password = "menas"
-  }
-  numThreads = 10
-}
